Move auth state fixture to conftest
